Ajustado estrutura dos pacotes, testes e Endpoint
