+++ src/main/java/br/com/blz/api/TestJavaApplication.java
@@ -1,4 +1,4 @@
-package br.com.blz;
+package br.com.blz.api;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
