+++ src/test/java/br/com/blz/testjava/TestJavaApplicationTests.java
@@ -2,6 +2,8 @@
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.util.Optional;
+
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,10 +13,10 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 import org.springframework.test.context.junit4.SpringRunner;
 
-import br.com.blz.models.InventoryEntity;
-import br.com.blz.models.ProductEntity;
-import br.com.blz.models.WarehouseEntity;
-import br.com.blz.repository.ProductRepository;
+import br.com.blz.api.models.InventoryEntity;
+import br.com.blz.api.models.ProductEntity;
+import br.com.blz.api.models.WarehouseEntity;
+import br.com.blz.api.repository.ProductRepository;
 
 @RunWith(SpringRunner.class)
 @DataJpaTest
@@ -26,17 +28,19 @@ public class TestJavaApplicationTests {
 
 	@Autowired
 	ProductRepository productDAO;
-
+	
 	@Test
-	public void should_find_products_if_repository_isNot_empty() {
+	public void should_find_repository_empty() {
+		productDAO.deleteAll();
 		Iterable<ProductEntity> products = productDAO.findAll();
-		assertThat(products).isNotEmpty();
+		assertThat(products).isNullOrEmpty();
 	}
+	
 	
 	@Test
 	  public void should_store_a_product() {
 		
-		//SETA WAREHOUSE 1
+				//SETA WAREHOUSE 1
 				WarehouseEntity warehouseEntity_x1 = new WarehouseEntity();
 				warehouseEntity_x1.setLocality("SP");
 				warehouseEntity_x1.setQuantity(12);
@@ -65,5 +69,20 @@ public class TestJavaApplicationTests {
 	 
 	    assertThat(productEntityNew).hasFieldOrPropertyWithValue("sku", new Long("99999"));	    
 	  }
+	
+	@Test
+	public void should_find_products_if_repository_isNot_empty() {
+		should_store_a_product();
+		Iterable<ProductEntity> products = productDAO.findAll();
+		assertThat(products).isNotEmpty();
+	}
+	
+	@Test
+	public void should_find_a_product() {
+		should_store_a_product();
+		long sku = 99999;
+		ProductEntity productEntityFind = productDAO.findById(sku);
+		assertThat(productEntityFind).hasFieldOrPropertyWithValue("sku", sku);
+	}
 
 }
